--- src/State/GifViewer.ts
@@ -1,7 +1,0 @@
-export class GifViewerState {
-  public srcList: string[] = [];
-  public delay: number = 0;
-  public repeat: number = 0;
-  public width: number = 0;
-  public height: number = 0;
-}
